deleted HTTP and refactored folder structure
