+++ db.ts
@@ -1,6 +1,6 @@
 import mongoose from 'mongoose';
 
-import config from '../../config';
+import config from './config';
 
 const MongoConnect = async () => {
     if (mongoose.connections[0].readyState) return;
