+++ models/ChatbotMessage.ts
